+++ go.mod
@@ -97,7 +97,6 @@ require (
 	github.com/nats-io/nats.go v1.10.0
 	github.com/newrelic/newrelic-telemetry-sdk-go v0.5.1
 	github.com/nsqio/go-nsq v1.0.8
-	github.com/ohler55/ojg v1.10.0
 	github.com/openconfig/gnmi v0.0.0-20180912164834-33a1865c3029
 	github.com/openzipkin/zipkin-go-opentracing v0.3.4
 	github.com/pkg/errors v0.9.1
