+++ plugins/parsers/enhancedjson/example.json
@@ -1,0 +1,53 @@
+{
+    "store": {
+        "book": [
+            {
+                "category": "reference",
+                "author": "Nigel Rees",
+                "title": "Sayings of the Century",
+                "price": 8.95
+            },
+            {
+                "category": "fiction",
+                "author": "Evelyn Waugh",
+                "title": "Sword of Honour",
+                "price": 12.99
+            },
+            {
+                "category": "fiction",
+                "author": "Herman Melville",
+                "title": "Moby Dick",
+                "isbn": "0-553-21311-3",
+                "price": 8.99,
+                "purchases": [
+                    "katy",
+                    "koala"
+                ]
+            },
+            {
+                "category": "fiction",
+                "author": "J. R. R. Tolkien",
+                "title": "The Lord of the Rings",
+                "isbn": "0-395-19395-8",
+                "price": 22.99,
+                "purchases": [
+                    "steve",
+                    "bob"
+                ]
+            }
+        ],
+        "decor": {
+            "color": "red",
+            "number-of-pillows": 19.95
+        }
+    }
+}
+
+
+node: book->purches->katy
+node: book->title->moby dick
+
+purchases title="Moby Dick",store_book_1_purchaser="katy"
+purchases title="Moby Dick",Purchaser="koala"
+purchases title="The Lord of The Rings",Purchaser="steve"
+purchases title="The Lord of The Rings",Purchaser="bob"
