+++ plugins/parsers/enhancedjson/testdata/current_weather/expected.out
@@ -1,0 +1,1 @@
+file,host=docker windspeed=1.5,timezone="Mountain View"
