+++ plugins/parsers/enhancedjson/testdata/current_weather/telegraf.conf
@@ -1,0 +1,19 @@
+[agent]
+hostname = "docker"
+
+# Parse usual values from current_weather JSON payload: https://openweathermap.org/current
+[[inputs.file]]
+files = ["./plugins/parsers/enhancedjson/testdata/current_weather/input.json"]
+data_format = "enhancedjson"
+        [[inputs.file.enhancedjson]]
+            [[inputs.file.enhancedjson.basic_fields]]
+            name = "timezone"
+            query = "name"
+            type = "string"
+            [[inputs.file.enhancedjson.basic_fields]]
+            name = "windspeed"
+            query = "wind.speed"
+            type = "float"
+
+[[outputs.file]]
+files = ["/tmp/current_weather.out"]
