+++ plugins/parsers/enhancedjson/testdata/types/telegraf.conf
@@ -3,96 +3,96 @@ hostname = "docker"
 
 # Parse String types from JSON
 [[inputs.file]]
-files = ["./plugins/parsers/jsonpath/testdata/types/input.json"]
-data_format = "jsonpath"
-        [[inputs.file.jsonpath]]
-            [[inputs.file.jsonpath.fields]]
+files = ["./plugins/parsers/enhancedjson/testdata/types/input.json"]
+data_format = "enhancedjson"
+        [[inputs.file.enhancedjson]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "explicitstringtypeName"
             query = "explicitstringtype"
             type = "string"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "defaultstringtypeName"
             query = "defaultstringtype"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertbooltostringName"
             query = "convertbooltostring"
             type = "string"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertinttostringName"
             query = "convertinttostring"
             type = "string"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertfloattostringName"
             query = "convertfloattostring"
             type = "string"
 
 # Parse int types from JSON
 [[inputs.file]]
-files = ["./plugins/parsers/jsonpath/testdata/types/input.json"]
-data_format = "jsonpath"
-        [[inputs.file.jsonpath]]
-            [[inputs.file.jsonpath.fields]]
+files = ["./plugins/parsers/enhancedjson/testdata/types/input.json"]
+data_format = "enhancedjson"
+        [[inputs.file.enhancedjson]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "explicitinttypeName"
             query = "explicitinttype"
             type = "int"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "defaultinttypeName"
             query = "defaultinttype"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertfloatointName"
             query = "convertfloatoint"
             type = "int"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertstringtointName"
             query = "convertstringtoint"
             type = "int"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertbooltointName"
             query = "convertbooltoint"
             type = "int"
 
 # Parse float types from JSON
 [[inputs.file]]
-files = ["./plugins/parsers/jsonpath/testdata/types/input.json"]
-data_format = "jsonpath"
-        [[inputs.file.jsonpath]]
-            [[inputs.file.jsonpath.fields]]
+files = ["./plugins/parsers/enhancedjson/testdata/types/input.json"]
+data_format = "enhancedjson"
+        [[inputs.file.enhancedjson]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "explicitfloattypeName"
             query = "explicitfloattype"
             type = "float"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "defaultfloattypeName"
             query = "defaultfloattype"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertintotfloatName"
             query = "convertintotfloat"
             type = "float"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertstringtofloatName"
             query = "convertstringtofloat"
             type = "float"
 
 # Parse bool types from JSON
 [[inputs.file]]
-files = ["./plugins/parsers/jsonpath/testdata/types/input.json"]
-data_format = "jsonpath"
-        [[inputs.file.jsonpath]]
-            [[inputs.file.jsonpath.fields]]
+files = ["./plugins/parsers/enhancedjson/testdata/types/input.json"]
+data_format = "enhancedjson"
+        [[inputs.file.enhancedjson]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "explicitbooltypeName"
             query = "explicitbooltype"
             type = "bool"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "defaultbooltypeName"
             query = "defaultbooltype"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertinttoboolName"
             query = "convertinttobool"
             type = "bool"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertstringtoboolName"
             query = "convertstringtobool"
             type = "bool"
-            [[inputs.file.jsonpath.fields]]
+            [[inputs.file.enhancedjson.basic_fields]]
             name = "convertintstringtoboolName"
             query = "convertintstringtobool"
             type = "bool"
