--- plugins/parsers/jsonpath/README.md
@@ -1,6 +1,0 @@
-# JSONPath
-
-THIS PARSER IS STILL A WORK IN PROGRESS
-
-The JSONPath data format parser parses a JSON string into metric fields using [JSONPath][https://goessner.net/articles/JsonPath/] expressions. For supported
-JSONPath query expressions check out [this article][https://goessner.net/articles/JsonPath/]. Examples for how this parser works can be found under the `testdata` directory, there you can find sample JSON, Telegraf configs, and expected metrics.
